web_api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Authlib==1.3.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
requests>=2.31.0

# GCP dependencies for TEE/Confidential Computing
google-cloud-compute>=1.15.0
google-cloud-storage>=2.10.0
google-cloud-kms>=2.19.0
google-auth>=2.23.0
google-api-python-client>=2.108.0
cryptography>=41.0.0
PyJWT>=2.8.0

